// src/app/services/core/language/en/messages/user.interfaces.ts
import { MessageInterface } from '../../langauge.interfaces';

export interface UserMessagesInterface {
	ConfirmAccountSuccess: MessageInterface;
	ConfirmAccountError: MessageInterface;
	ConfirmEmailSuccess: MessageInterface;
	ConfirmEmailError: MessageInterface;
	EmailUpdateSuccess: MessageInterface;
	EmailUpdateError: MessageInterface;
	NameUpdateSuccess: MessageInterface;
	NameUpdateError: MessageInterface;
	PasswordUpdateSuccess: MessageInterface;
	PasswordUpdateError: MessageInterface;
}